requirements.txt
ipykernel
python-dotenv
google-api-python-client
kafka-python
pyspark
faker